Same class of HTML wrapper element
When creating a custom Notification with the `NotificationService.html()` method, the `sn-content` class was added around the entire HTML code one passed to the method. This should have been the `sn-html` class, in my opinion.

When creating the notification from an Angular template, the class `sn-html` is added, which looks correct to me. I changed the code so that the `sn-html` class is assigned in both cases.